+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/manager/EventCatchupProcessor.java
@@ -1,7 +1,7 @@
 package uk.gov.justice.services.event.sourcing.subscription.manager;
 
 import static java.lang.String.format;
-import static javax.transaction.Transactional.TxType.REQUIRES_NEW;
+import static javax.transaction.Transactional.TxType.NOT_SUPPORTED;
 
 import uk.gov.justice.services.event.source.subscriptions.repository.jdbc.SubscriptionsRepository;
 import uk.gov.justice.services.eventsourcing.source.core.EventSource;
@@ -29,7 +29,7 @@ public class EventCatchupProcessor {
         this.logger = logger;
     }
 
-    @Transactional(REQUIRES_NEW)
+    @Transactional(NOT_SUPPORTED)
     public void performEventCatchup(final Subscription subscription) {
 
         final EventSource eventSource = eventSourceProvider.getEventSource(subscription.getEventSourceName());